refactored twittercontroller to twitterCrawler
